requirements.txt
apache-airflow-providers-microsoft-azure>=1.2.0
